+++ app/Models/Avaliacao.php
@@ -8,7 +8,7 @@ use Illuminate\Database\Eloquent\Model;
 class Avaliacao extends Model
 {
     use HasFactory;
-
+    protected $table = 'avaliacao';
     public function aluno()
     {
         return $this->belongsTo(Aluno::class);
