+++ public/css/style.css
@@ -86,7 +86,7 @@ footer {
     background-color: #C7A6DB;
     color: black;
     padding: 5px;
-    position: sticky; /* Alterado para 'sticky' */
+    position: fixed;
     bottom: 0;
     width: 100%;
     top: auto; /* Define 'top' como 'auto' para iniciar o 'sticky' quando o footer atingir o topo da página */
